[#318] Fix Sonar Cloud code smell - split big file into smaller files
