+++ CMakeLists.txt
@@ -44,6 +44,8 @@ idf_component_register(
         src/http_server_handle_req_post_auth.h
         src/http_server_mutex.c
         src/http_server_mutex.h
+        src/http_server_netconn_resp.c
+        src/http_server_netconn_resp.h
         src/http_server_resp.c
         src/sta_ip_safe.c
         src/sta_ip_safe.h
